\name{ContaminatedMixt-package}
\alias{ContaminatedMixt-package}
\alias{ContaminatedMixt}
\docType{package}
\title{ContaminatedMixt - Parsimonious Mixtures of Contaminated Normal Distributions}
\description{
This package allows to fit, according to the expectation-conditional maximization algorithm, the 14 parsimonious mixtures of multivariate contaminated normal distributions, with eigen-decomposed scale matrices, introduced by Punzo and McNicholas (2015).
Model-based clustering and classification scenarios are implemented.
Likelihood-based model selection criteria can be adopted to select the parsimonious model and the number of groups.
}
\details{
\tabular{ll}{
Package: \tab ContaminatedMixt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-10\cr
License: \tab GNU-2\cr
}
}
\references{
Punzo, A. and McNicholas, P. D. (2015). Parsimonious mixtures of contaminated Gaussian distributions with application to allometric studies. \emph{arXiv.org} e-print \bold{1305.4669}, available at: \url{http://arxiv.org/abs/1305.4669}.
}
\author{
Antonio Punzo, Angelo Mazza, Paul D. McNicholas 

Maintainer: Angelo Mazza <a.mazza@unict.it>
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{CNmixt}}
}