\name{ContaminatedMixt-package}
\alias{ContaminatedMixt-package}
\alias{ContaminatedMixt}
\docType{package}
\title{ContaminatedMixt - Parsimonious Mixtures of Contaminated Normal Distributions}
\description{
This package allows to fit, according to the expectation-conditional maximization algorithm, the 14 parsimonious mixtures of multivariate contaminated normal distributions, with eigen-decomposed scale matrices, introduced by Punzo and McNicholas (2016).
Model-based clustering and classification scenarios are implemented.
Likelihood-based model selection criteria can be adopted to select the parsimonious model and the number of groups.
}
\details{
\tabular{ll}{
Package: \tab ContaminatedMixt\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-09-02\cr
License: \tab GNU-2\cr
}
}
\references{
Punzo A., Mazza A. and McNicholas P. D. (2018). \pkg{ContaminatedMixt}: An R Package for Fitting Parsimonious Mixtures of Multivariate Contaminated Normal Distributions. \emph{Journal of Statistical Software}.

Punzo A. and McNicholas P. D. (2016). Parsimonious mixtures of multivariate contaminated normal distributions. \emph{Biometrical Journal}, \bold{58}(6), 1506--1537.
}
\author{
Antonio Punzo, Angelo Mazza, Paul D. McNicholas 

Maintainer: Angelo Mazza <a.mazza@unict.it>
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{CNmixt}}
}